Task3 is completed
Added the code to the third task on HTML and CSS. I find more comfortable placing the code here, don't you think so? GDrive is not appropriate for such purposes)) Two files: HTML with the structure and CSS with the design. You can find implementation and usage of borders and classes there!